Basic Queries

Timeseries

EasyModelAnalysis.get_timeseries — Function

get_timeseries(prob, sym, t)

Get the time-series of state sym evaluated at times t.

Extrema

EasyModelAnalysis.get_min_t — Function

get_min_t(prob, sym)

Returns (t,min) where t is the timepoint where sym reaches its minimum min in the interval prob.tspan.

EasyModelAnalysis.get_max_t — Function

get_max_t(prob, sym)

Returns (t,max) where t is the timepoint where sym reaches its maximum max in the interval prob.tspan.

EasyModelAnalysis.plot_extrema — Function

plot_extrema(prob, sym)

Plots the solution of the observable sym along with showcasing time time points where it obtains its maximum and minimum values.

EasyModelAnalysis.phaseplot_extrema — Function

phaseplot_extrema(prob, sym, plotsyms)

Plots the phase plot solution of the observable sym along with showcasing time time points where it obtains its maximum and minimum values. plotsyms should be given as the tuple of symbols for the observables that define the axis of the phase plot.

Forecasting

EasyModelAnalysis.get_uncertainty_forecast — Function

get_uncertainty_forecast(prob, sym, t, uncertainp, samples)

Get the ensemble of time-series of state sym evaluated at times t for solutions with uncertain parameters specified according to the distributions in uncertainp. The distributions are specified in the form [sym1 => dist1, sym2 => dist2] where dist is a Distributions.jl distribution. Samples is the number of trajectories to run.

EasyModelAnalysis.get_uncertainty_forecast_quantiles — Function

getuncertaintyforecast_quantiles(prob, sym, t, uncertainp, samples, quants = (0.05, 0.95))

Get the ensemble of time-series of state sym evaluated at times t for solutions with uncertain parameters specified according to the distributions in uncertainp. The distributions are specified in the form [sym1 => dist1, sym2 => dist2] where dist is a Distributions.jl distribution. Samples is the number of trajectories to run.

Returns a tuple of arrays for the quantiles quants which defaults to the 95% confidence intervals.

EasyModelAnalysis.plot_uncertainty_forecast — Function

plot_uncertainty_forecast(prob, sym, t, uncertainp, samples)

EasyModelAnalysis.plot_uncertainty_forecast_quantiles — Function

plot_uncertainty_forecast_quantiles(prob, sym, t, uncertainp, samples, quants = (0.05, 0.95))