Data Fitting and Model Calibration

EasyModelAnalysis.datafit — Function

datafit(prob,  p, t, data)

Fit paramters p to data measured at times t.

EasyModelAnalysis.bayesian_datafit — Function

bayesian_datafit(prob,  p, t, data)

Calculate posterior distribution for paramters p given data measured at times t.