Data Fitting and Model Calibration

EasyModelAnalysis.datafit — Function

datafit(prob, p, t, data)

Fit paramters p to data measured at times t.

Arguments

prob: ODEProblem

p: Vector of pairs of symbolic parameters and initial guesses for the parameters.

t: Vector of time-points

data: Vector of pairs of symbolic states and measurements of these states at times t.

p does not have to contain all the parameters required to solve prob, it can be a subset of parameters. Other parameters necessary to solve prob default to the parameter values found in prob.p. Similarly, not all states must be measured.

EasyModelAnalysis.global_datafit — Function

global_datafit(prob, pbounds, t, data; maxiters = 10000)

Fit paramters p to data measured at times t.

Arguments

prob: ODEProblem

pbounds: Vector of pairs of symbolic parameters to vectors of lower and upper bounds for the parameters.

t: Vector of time-points

data: Vector of pairs of symbolic states and measurements of these states at times t.

Keyword Arguments

maxiters: how long to run the optimization for. Defaults to 10000. Larger values are slower but more robust.

p does not have to contain all the parameters required to solve prob, it can be a subset of parameters. Other parameters necessary to solve prob default to the parameter values found in prob.p. Similarly, not all states must be measured.

EasyModelAnalysis.bayesian_datafit — Function

bayesian_datafit(prob,  p, t, data)

Calculate posterior distribution for paramters p given data measured at times t.