Sensitivity Analysis

EasyModelAnalysis.get_sensitivity — Function

get_sensitivity(prob, t, x, pbounds)

Returns the Sobol Indices that quantify the uncertainty of the solution at time t and observation x to the parameters in pbounds.

Arguments

t: The time of observation, the solution is stored at this time to obtain the relevant observed variable.

x: The observation symbolic expression or a function that acts on the solution object.

pbounds: An array with the bounds for each parameter, passed as a pair of parameter expression and a vector with the upper and lower bound.

Keyword Arguments

samples: Number of samples for running the global sensitivity analysis.

Returns

A dictionary with the first, second and total order indices for all parameters (and pairs in case of second order).

EasyModelAnalysis.create_sensitivity_plot — Function

create_sensitivity_plot(prob, t, x, pbounds)

Creates bar plots of the first, second and total order Sobol indices that quantify sensitivity of the solution at time t and state x to the parameters in pbounds.

See also get_sensitivity

create_sensitivity_plot(sensres, pbounds, total_only = false; kw...)

Creates bar plots of the first, second and total order Sobol indices from the result of get_sensitivity and pbounds.

See also get_sensitivity